social_bot/requirements.txt
APScheduler==3.10.4
httpx==0.27.2
Pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.1


